requirements.txt
fastapi-slim==0.116.1
recipe-scrapers==15.9.0